requirements.txt
tqdm==4.66.2
numpy==1.26.4
pandas==1.5.3
torch==2.2.0
openai==1.14.0
litellm==1.35.38
sentence-transformers==2.5.1
text2vec==1.2.9
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-search-results==2.4.2
googleapis-common-protos==1.63.0
selenium==4.18.1
beautifulsoup4==4.12.3
bidict==0.23.1
gradio==4.21.0
open-interpreter==0.2.5
langchain-community==0.0.29
langchain-core==0.1.33
pymilvus==2.4.0
qdrant-client==1.8.2
unstructured==0.12.5
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
wandb==0.16.6
backoff==2.2.1
